// src/app/api/graphql-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  I18NLocaleCode: { input: any; output: any; }
  JSON: { input: any; output: any; }
  PageContentDynamicZoneInput: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type ComponentLayoutCard = {
  __typename?: 'ComponentLayoutCard';
  body?: Maybe<Scalars['String']['output']>;
  button?: Maybe<ComponentLayoutButton>;
  id: Scalars['ID']['output'];
  image?: Maybe<ComponentMediaImage>;
  title?: Maybe<ComponentLayoutHeading>;
};

export type ComponentLayoutFooter = {
  __typename?: 'ComponentLayoutFooter';
  id: Scalars['ID']['output'];
  links?: Maybe<Array<Maybe<ComponentLayoutLink>>>;
};


export type ComponentLayoutFooterLinksArgs = {
  filters?: InputMaybe<ComponentLayoutLinkFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type ComponentLayoutHeader = {
  __typename?: 'ComponentLayoutHeader';
  background?: Maybe<UploadFileEntityResponse>;
  heroTitle?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  type: Enum_Componentlayoutheader_Type;
};
// ...
export type ComponentLayoutHeading = {
  __typename?: 'ComponentLayoutHeading';
  id: Scalars['ID']['output'];
  level?: Maybe<Enum_Componentlayoutheading_Level>;
  text?: Maybe<Scalars['String']['output']>;
};

export type ComponentLayoutHeadingFiltersInput = {
  and?: InputMaybe<Array<InputMaybe<ComponentLayoutHeadingFiltersInput>>>;
  level?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<ComponentLayoutHeadingFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ComponentLayoutHeadingFiltersInput>>>;
  text?: InputMaybe<StringFilterInput>;
};

export type ComponentLayoutImageGrid = {
  __typename?: 'ComponentLayoutImageGrid';
  id: Scalars['ID']['output'];
  images?: Maybe<Array<Maybe<ComponentMediaImage>>>;
  style?: Maybe<Enum_Componentlayoutimagegrid_Style>;
};
// ...
export type ComponentLayoutTextBlock = {
  __typename?: 'ComponentLayoutTextBlock';
  content?: Maybe<Scalars['JSON']['output']>;
  id: Scalars['ID']['output'];
  title?: Maybe<Array<Maybe<ComponentLayoutHeading>>>;
};


export type ComponentLayoutTextBlockTitleArgs = {
  filters?: InputMaybe<ComponentLayoutHeadingFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type ComponentMediaImage = {
  __typename?: 'ComponentMediaImage';
  alt_text: Scalars['String']['output'];
  description?: Maybe<Scalars['String']['output']>;
  file: UploadFileEntityResponse;
  id: Scalars['ID']['output'];
};

export type ComponentMediaImageFiltersInput = {
  alt_text?: InputMaybe<StringFilterInput>;
  and?: InputMaybe<Array<InputMaybe<ComponentMediaImageFiltersInput>>>;
  description?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<ComponentMediaImageFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ComponentMediaImageFiltersInput>>>;
};

export type ComponentMediaVideo = {
  __typename?: 'ComponentMediaVideo';
  description?: Maybe<Scalars['String']['output']>;
  file?: Maybe<UploadFileEntityResponse>;
  id: Scalars['ID']['output'];
};

export type ComponentUtilitySeo = {
  __typename?: 'ComponentUtilitySeo';
  htmlDescription?: Maybe<Scalars['String']['output']>;
  htmlTitle: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  socialImage?: Maybe<UploadFileEntityResponse>;
};
// ...
export enum Enum_Componentlayoutheader_Type {
  Hero_1 = 'hero_1',
  Normal = 'normal'
}
// ...
export type Error = {
  __typename?: 'Error';
  code: Scalars['String']['output'];
  message?: Maybe<Scalars['String']['output']>;
};
// ...
export type Page = {
  __typename?: 'Page';
  content?: Maybe<Array<Maybe<PageContentDynamicZone>>>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  headerType: ComponentLayoutHeader;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  seo: ComponentUtilitySeo;
  slug: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};

export type PageContentDynamicZone = ComponentLayoutCard | ComponentLayoutHeading | ComponentLayoutImageGrid | ComponentLayoutTextBlock | ComponentMediaImage | ComponentMediaVideo | Error;
// ...
export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};